Removing tags from org-roam-node-display-template
Tag display with ido doesn't work as expected, buffer is filled with
newline chars and org-roam-node-find fails to autocomplete downloading
a file.

Problem seems to be tag management in roam v2. I'm not leveraging tags
and solely relying on backlinks for organising my notes. Meaning, I
can only display titles and be fine with it.